feat: add logic to calculate the cart subtotal in the view model
