+++ react-client/src/components/Membership/CartFooter.tsx
@@ -2,15 +2,10 @@ import { Button, Text } from "../design-system";
 
 type CartFooterProps = {
   onClick?: () => void;
-  shouldProceedToCheckout?: boolean;
   subtotal: string;
 };
 
-export const CartFooter = ({
-  onClick,
-  shouldProceedToCheckout,
-  subtotal,
-}: CartFooterProps) => {
+export const CartFooter = ({ onClick, subtotal }: CartFooterProps) => {
   return (
     <>
       <div className="flex w-full justify-between">
@@ -27,9 +22,7 @@ export const CartFooter = ({
         </Text>
       </div>
 
-      <Button onClick={onClick} disabled={shouldProceedToCheckout}>
-        Proceed to checkout
-      </Button>
+      <Button onClick={onClick}>Proceed to checkout</Button>
     </>
   );
 };
